% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pajam-heatmap.R
\name{list2im_opt}
\alias{list2im_opt}
\title{List to incidence matrix}
\usage{
list2im_opt(setlist, ...)
}
\description{
List to incidence matrix
}
\details{
This function takes a list of \code{character} vectors and returns
a \code{numeric} \code{matrix} whose rownames are entries from the
\code{character} vectors, colnames are \code{names} from the \code{setlist},
and whose values are \code{0} or \code{1}.
}
\examples{
setlist <- list(setA=LETTERS[1:10], setB=LETTERS[7:14]);
list2im_opt(setlist);

}
\concept{pajam list}
