% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pajam-shiny.R
\name{pajam_shiny_ui}
\alias{pajam_shiny_ui}
\title{R-shiny UI for Protein Atlas visualization}
\usage{
pajam_shiny_ui(...)
}
\arguments{
\item{...}{additional arguments are ignored.}
}
\description{
R-shiny UI for Protein Atlas visualization
}
\seealso{
Other pajam shiny: 
\code{\link{launch_pajam}()},
\code{\link{pajam_shiny_server}()}
}
\concept{pajam shiny}
