% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pajam-data.R
\docType{data}
\name{proteinatlas_genesets_fdb11}
\alias{proteinatlas_genesets_fdb11}
\title{Protein Atlas useful genesets for visualization}
\format{\code{list} of \verb{character vector}, whose values match
\code{rownames(proteinatlas_expr_fdb11)}.}
\usage{
proteinatlas_genesets_fdb11
}
\description{
Protein Atlas useful genesets for visualization
}
\details{
This object includes a list of genesets in the form of
character vectors of rownames aligned with
\code{proteinatlas_expr_fdb11}. The genesets are some useful
annotations for visualization that include:
\itemize{
\item \code{"Plasma"} - found in circulating blood plasma
\item \code{"secreted_proteins"} - secreted proteins
\item \code{"membrane_proteins"} - membrane-bound proteins
\item \code{"FDA"} - gene targets of FDA-approved drugs
\item \code{"NOT_membrane_secreted"} - non-membrane and non-secreted proteins (intracellular)
\item \code{"TFs"} - transcription factors
\item \code{"HKs"} - housekeeper genes
\item \code{"rna_lymphoid"} - RNA-lymphoid associated genes
}

The annotations are not intended to be fully exhaustive,
but are intended to represent some useful annotations
not necessarily associated directly with the biological
function of each gene/protein. For example, it may be useful
to see whether most up-regulated proteins are secreted or
intracellular, or perhaps most up-regulated gene transcripts
are transcription factors and are not secreted.
}
\examples{
use_im <- c("secreted_proteins",
   "membrane_proteins",
   "NOT_membrane_secreted",
   "TFs");
proteinatlas_im <- list2im_opt(proteinatlas_genesets_fdb11[use_im]);
test_genes <- c("DKK1","DKK4","CXCL12","IL6R","MET",
   "HK2","FTL","FTH1","STAT1","STAT3","CDKN1B");

proteinatlas_heatmap(genes=test_genes,
   type="Blood",
   centered=TRUE,
   gene_im=proteinatlas_im);

proteinatlas_heatmap(genes=test_genes,
   type="Blood",
   centered=TRUE,
   column_filter=4,
   row_cex=0.8,
   column_names_max_height=grid::unit(12, "cm"),
   gene_im=proteinatlas_im);

}
\seealso{
Other pajam data: 
\code{\link{proteinatlas_expr_fdb11}}
}
\concept{pajam data}
\keyword{datasets}
