% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pajam-heatmap.R
\name{proteinatlas_heatmap}
\alias{proteinatlas_heatmap}
\title{proteinatlas.org expression heatmap}
\usage{
proteinatlas_heatmap(
  expr = proteinatlas_expr_fdb11,
  genes = NULL,
  samples = NULL,
  type = c("all", "Tissue", "Cell", "Blood", "Brain"),
  row_cex = 1,
  column_cex = 1,
  ramp = "Reds",
  lens = 0,
  color_ceiling = NULL,
  cluster_columns = FALSE,
  cluster_rows = FALSE,
  column_split = NULL,
  centered = FALSE,
  row_filter = 0,
  column_filter = 0,
  controlSamples = NULL,
  gene_names = FALSE,
  gene_im = NULL,
  fill_missing = TRUE,
  border = TRUE,
  useCenterGroups = TRUE,
  trim_columns = FALSE,
  rowStatsFunc = matrixStats::rowMins,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{expr}{\code{numeric} matrix containing gene rows, and
biological sample columns.}

\item{genes}{\code{character} vector of genes to display. These values
are matched case-insensitive to the start of values in
\code{rownames(expr)}, in order to facilitate pattern-matching.
Specifically, the input is passed to
\code{jamba::provigrep(c(genes), rownames(expr))} which iteratively
matches each input \code{genes} using case-insensitive \code{grep()}.}

\item{type}{\code{character} vector of one or more expression types
to include, where \code{"all"} includes all columns in \code{expr}.}

\item{row_cex, column_cex}{\code{numeric} value used to adjust row and
column text size, where \code{1} is the default size. Text size is
auto-scaled based upon the number of rows and columns being
displayed; these values adjust the auto-scaled font size.}

\item{ramp}{\code{character} name of a color ramp, or a \code{character}
vector of R colors, used to create a color ramp.}

\item{lens}{\code{numeric} value used to adjust the intensity of
the color ramp, where values > 0 make colors change earlier,
and values < 0 make colors change later.}

\item{color_ceiling}{\code{numeric} value indicating the maximum used
for the color gradient, using units as displayed on the heatmap
(exponentiated.) If \code{NULL} then the maximum value is used.}

\item{cluster_columns, cluster_rows}{\code{logical} indicating whether to
cluster columns and rows, respectively.}

\item{centered}{\code{logical} indicating whether data should be centered
by the median expression across all samples. Note that median
expression in many cases is zero, for genes not widely expressed
across all samples being displayed.}

\item{column_filter, row_filter}{numeric value which hides columns or
rows when the column or row maximum expression is not at or above
this numeric threshold. The value filtered is the expression value
indicated on the heatmap, the normal expression value, not
log2-transformed.}

\item{controlSamples}{\code{character} vector, or \code{NULL}, indicating
specific samples to use when centering data by row. When
\code{controlSamples=NULL}, the default is to use all Tissue samples.}

\item{gene_names}{\code{logical} indicating whether to display full gene
names, provided using \code{genejam::freshenGenes()}.}

\item{gene_im}{\code{matrix} intended to be an incidence matrix, whose
values are \code{0} and \code{1}, with gene symbols as rownames, and various
columns indicating different annotations. When a rowname in the
heatmap is not found in the incidence matrix,
it subtitutes \code{0} to fill the empty space. See examples to see how
to use the \code{proteinatlas_genesets_fdb11} data.}

\item{fill_missing}{\code{logical} indicating whether the input \code{genes} should
bo used as-is with no pattern matching, and by substituting \code{0}
for any missing entries. This argument is useful when trying to
align this heatmap with another existing \code{Heatmap} object, where
the rows must be exactly aligned.}

\item{border}{\code{logical} indicating whether to draw a heatmap border.}

\item{useCenterGroups}{\code{logical} used when \code{centered=TRUE}, and when
\code{controlSamples} are provided, will center each sample type
independently, instead of centering all samples versus \code{"Tissue"}
type. We decided by default to center all samples versus \code{"Tissue"}
by default, because some immune-specific genes are highly expressed
in all \code{"Blood"} samples but not \code{"Tissue"}, and it was visually
confusing.}

\item{trim_columns}{\code{logical} indicating whether to trim the colnames
to remove the sample type suffix.}

\item{rowStatsFunc}{\code{function} used to define the per-row value used
when \code{centered=TRUE}, passed to \code{jamma::centerGeneData_new()}.
The default is to use the \code{matrixStats::rowMins()} which returns
the minimum expression value per row.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are passed to \code{ComplexHeatmap::Heatmap()}
for other customizations.}
}
\value{
\code{Heatmap} produced by \code{ComplexHeatmap::Heatmap()}.
}
\description{
proteinatlas.org expression heatmap
}
\details{
This function takes proteinatlas expression data \code{expr},
and creates a heatmap of expression using a subset of
genes provided as \code{genes}.

By default, columns in \code{expr} are split by type, where
\code{colnames(expr)} are expected to have suffix \code{" - Type"}
at the end of each column name. If the columns cannot be
split accordingly, then all columns are assigned one
split name \code{"Expression"}.

To customize data for individual samples, the \code{expr} data
should be filtered before calling this function.

By default, the row-centering method centers each row by
the row minimum expression using only \code{"Tissue"} samples,
so that expression will be displayed relative to the
lowest tissue expression. An optional set of control samples
can be provided with argument \code{controlSamples}.
}
\examples{
test_genes <- c("DKK1","DKK4","CXCL12","IL6R","MET",
   "HK2","FTL","FTH1","STAT1","STAT3","CDKN1B");
proteinatlas_heatmap(genes=test_genes,
   type="Blood",
   centered=FALSE,
   cluster_rows=TRUE,
   cluster_columns=TRUE,
   row_filter=2)

# use proteinatlas_genesets_fdb11
use_im <- c("secreted_proteins",
   "membrane_proteins",
   "NOT_membrane_secreted",
   "TFs");
proteinatlas_im <- list2im_opt(proteinatlas_genesets_fdb11[use_im]);
test_genes <- c("DKK1","DKK4","CXCL12","IL6R","MET",
   "HK2","FTL","FTH1","STAT1","STAT3","CDKN1B");
rowGroupMeans <- jamba::rowGroupMeans;
proteinatlas_heatmap(genes=test_genes,
   type="Blood",
   centered=TRUE,
   gene_im=proteinatlas_im);

}
\concept{pajam heatmap}
