% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pajam-shiny.R
\name{launch_pajam}
\alias{launch_pajam}
\title{Launch pajam shiny app}
\usage{
launch_pajam(
  ...,
  width = 2200,
  port = 8888,
  host = "0.0.0.0",
  options = list(width = width, host = host, port = port)
)
}
\description{
Launch pajam shiny app
}
