% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pajam-shiny.R
\name{pajam_shiny_server}
\alias{pajam_shiny_server}
\title{R-shiny UI for Protein Atlas visualization}
\usage{
pajam_shiny_server(input, output, session, shiny_env = new.env(), ...)
}
\description{
R-shiny UI for Protein Atlas visualization
}
